completed move of automation to have .h & .cpp file, did same with solenoids.h to prevent multiple definition error during build.
