+++ lib/Automation/Automation.cpp
@@ -5,11 +5,9 @@
 
 #include "Automation.h"
 
-
-#include <Arduino.h>
-#include <solenoids.h>
-
 using namespace std;
+
+namespace Automation {
 
 //-----------------------Variables-----------------------
 
@@ -196,3 +194,4 @@ using namespace std;
     return true;
   }
   
+} //Automation
